Start splitting capture-changes modal into wizard.
